some pep8, verbose mode working again, and actool changes

+++ munki_rebrand.py
@@ -66,7 +66,7 @@ MSC_APP = {'path': 'Applications/Managed Software Center.app/Contents/Resources'
 MS_APP = {'path': os.path.join(MSC_APP['path'], 'MunkiStatus.app/Contents/Resources'),
           'icon': ['MunkiStatus.icns', 'AppIcon.icns']}
 
-APPS = [ MSC_APP, MS_APP ]
+APPS = [MSC_APP, MS_APP]
 
 ICON_SIZES = [('16', '16x16'), ('32', '16x16@2x'),
               ('32', '32x32'), ('64', '32x32@2x'),
@@ -92,7 +92,8 @@ global verbose
 verbose = False
 tmp_dir = mkdtemp()
 
-#@atexit.register
+
+@atexit.register
 def cleanup():
     print "Cleaning up..."
     try:
@@ -103,7 +104,7 @@ def cleanup():
     print "Done."
 
 
-def run_cmd(cmd, ret=None, verbose=True):
+def run_cmd(cmd, ret=None):
     '''Runs a command passed in as a list. Can also be provided with a regex
     to search for in the output, returning the result'''
     proc = Popen(cmd, stdout=PIPE, stderr=PIPE)
@@ -123,6 +124,7 @@ def get_latest_munki_url():
     j = run_cmd(cmd, ret=True)
     api_result = json.loads(j)
     return api_result['assets'][0]['browser_download_url']
+
 
 def download_pkg(url, output):
     print "Downloading munkitools from %s..." % url
@@ -149,7 +151,7 @@ def expand_payload(payload, directory):
     '''Expands a pkg payload to a directory'''
     cmd = [DITTO]
     # Ditto verbose is to stderr :(
-    #if verbose:
+    # if verbose:
     #    cmd.extend(['-v'])
     cmd.extend(['-x', payload, directory])
     run_cmd(cmd)
@@ -200,6 +202,7 @@ def plist_to_binary(plist):
     cmd = [PLUTIL, '-convert', 'binary1', plist]
     run_cmd(cmd)
 
+
 def guess_encoding(string):
     encodings = ['ascii', 'utf-8', 'utf-16']
     for encoding in encodings:
@@ -209,7 +212,8 @@ def guess_encoding(string):
         except UnicodeError:
             continue
 
-def replace_strings(strings_file, code, appname, verbose=True):
+
+def replace_strings(strings_file, code, appname):
     '''Replaces localized app name in a .strings file with desired app name'''
     localized = APPNAME_LOCALIZED[code]
     if verbose:
@@ -221,8 +225,10 @@ def replace_strings(strings_file, code, appname, verbose=True):
         rawdata = f.read()
         enc = guess_encoding(rawdata)
 
-    with io.open(backup_file, 'w', encoding=enc) as fw, \
-         io.open(strings_file, 'r', encoding=enc) as fr:
+    # OK then, pep8...
+    with io.open(backup_file, 'w',
+                 encoding=enc) as fw, io.open(strings_file, 'r',
+                                              encoding=enc) as fr:
         for line in fr:
             # We want to only replace on the right hand side of any =
             # and we don't want to do it to a comment
@@ -244,7 +250,9 @@ def replace_nib(nib_file, code, appname):
                                                          appname)
     backup_file = '%s.bak' % nib_file
     plist_to_xml(nib_file)
-    with io.open(backup_file, 'w', encoding='utf-8') as fw,  io.open(nib_file, 'r', encoding='utf-8') as fr:
+    with io.open(backup_file, 'w',
+                 encoding='utf-8') as fw,  io.open(nib_file, 'r',
+                                                   encoding='utf-8') as fr:
         for line in fr:
             # Simpler than mucking about with plistlib
             line = line.replace(localized, appname.decode('utf-8'))
@@ -279,30 +287,36 @@ def convert_to_icns(png, output_dir, actool=""):
                      scale=scale)
         contents['images'].append(image)
     icns = os.path.join(icon_dir, 'AppIcon.icns')
+
     # Munki 3.6+ has an Assets.car which is compiled from the Assets.xcassets
     # to provide the AppIcon
     if actool:
         with open(os.path.join(iconset, 'Contents.json'), 'w') as f:
             json.dump(contents, f)
         cmd = [actool,
-            '--compile', icon_dir,
-            '--app-icon', 'AppIcon',
-            '--minimum-deployment-target', '10.8',
-            '--output-partial-info-plist', os.path.join(icon_dir, 'Info.plist'),
-            '--platform', 'macosx',
-            '--errors',
-            '--warnings',
-            xcassets]
+               '--compile', icon_dir,
+               '--app-icon', 'AppIcon',
+               '--minimum-deployment-target', '10.8',
+               '--output-partial-info-plist', os.path.join(icon_dir,
+                                                           'Info.plist'),
+               '--platform', 'macosx',
+               '--errors',
+               '--warnings',
+               xcassets]
         run_cmd(cmd)
     else:
         # Old behaviour for < 3.6
         cmd = [ICONUTIL, '-c', 'icns', iconset,
-           '-o', icns]
+               '-o', icns]
         run_cmd(cmd)
 
-    car = os.path.join(icon_dir, 'Assets.car')
+    carpath = os.path.join(icon_dir, 'Assets.car')
+    if os.path.isfile(carpath):
+        car = carpath
+    else:
+        car = None
 
-    return icns, car or None
+    return icns, car
 
 
 def sign_package(signing_id, pkg):
@@ -326,15 +340,14 @@ def main():
     p.add_argument('-a', '--appname', action='store',
                    required=True,
                    help="Your desired app name for Managed Software "
-                   "Center."),
+                   "Center.")
     p.add_argument('-k', '--pkg', action='store',
                    help="Prebuilt munkitools pkg to rebrand."),
     p.add_argument('-i', '--icon-file', action='store',
                    default=None,
                    help="Optional icon file to replace Managed Software "
-                   "Center's. Can be a .icns file or a 1024x1024 .png with "
-                   "alpha channel, in which case it will be converted to an "
-                   ".icns")
+                   "Center's. Should be a 1024x1024 .png with "
+                   "alpha channel")
     p.add_argument('-o', '--output-file', action='store',
                    default=None,
                    help="Optional base name for the customized pkg "
@@ -358,8 +371,7 @@ def main():
               "munki installer pkg!"
         sys.exit(1)
 
-
-
+    global verbose
     verbose = args.verbose
     outfilename = args.output_file or "munkitools"
 
@@ -374,10 +386,9 @@ def main():
             # Attempt to convert png to icns
             print "Converting .png file to .icns..."
             args.icon_file, car = convert_to_icns(args.icon_file,
-                                             tmp_dir,
-                                             actool=actool)
+                                                  tmp_dir,
+                                                  actool=actool)
     output = os.path.join(tmp_dir, 'munkitools.pkg')
-
 
     if not args.pkg:
         download_pkg(get_latest_munki_url(), output)
@@ -428,6 +439,7 @@ def main():
         shutil.rmtree(app_pkg)
 
         # Find the lproj directories in the apps' Resources dirs
+        print "Replacing app name with %s..." % args.appname
         for app in APPS:
             resources_dir = os.path.join(app_payload, app['path'])
             # Get a list of all the lproj dirs in each app's Resources dir
@@ -449,18 +461,19 @@ def main():
                     if os.path.isfile(
                          os.path.join(app_payload,
                                       os.path.join(app['path'], icon))):
-                        print (app['path'], icon)
                         found_icon = icon
                         break
                 icon_path = os.path.join(app['path'], found_icon)
                 dest = os.path.join(app_payload, icon_path)
-                print "Replacing icons with %s in %s..." % (args.icon_file,
-                                                            dest)
+                print "Replacing icons in %s with %s..." % (dest,
+                                                            args.icon_file)
                 shutil.copyfile(args.icon_file, dest)
                 if car:
                     car_path = os.path.join(app['path'], 'Assets.car')
                     dest = os.path.join(app_payload, car_path)
-                    shutil.copyfile(car, dest)
+                    if os.path.isfile(dest):
+                        shutil.copyfile(car, dest)
+                        print "Replacing icons in %s with %s..." % (dest,car)
 
         # Make a new root for the distribution product
         newroot = os.path.join(tmp_dir, 'newroot')
@@ -499,6 +512,7 @@ def main():
 
     else:
         print "Could not find munkitools pkg %s." % args.pkg
+
 
 if __name__ == '__main__':
     main()
